Move bounce animation to util class
